fix(cli): resolve interactive shell PR checks
Made-with: Cursor
